inserting exception in remove asset use case

+++ src/main/java/br/edu/ifsp/addthenewsoul/domain/usecases/asset/RemoveAssetUseCase.java
@@ -13,6 +13,10 @@ public class RemoveAssetUseCase {
 
     public void removesAsset (Asset asset) {
         if (asset == null) return;
+
+        Integer id = asset.getId();
+        if (assetDAO.findById(id).isEmpty())
+            throw new EntityAlreadyExistsException("This resource ID does not exist");
         assetDAO.delete(asset.getId());
     }
 }
